test: share one mock OpenAI upstream across the doc-snippet fixtures
